Moved some files around and removed `module-alias` and introduced `babel` dev-dependency
