+++ .babelrc
@@ -1,0 +1,19 @@
+{
+  "presets": [
+    "@babel/preset-env",
+    "@babel/preset-typescript",
+    "minify"
+  ],
+  "plugins": [
+    ["babel-plugin-module-resolver", {
+      "root": ["./src"],
+      "alias": {
+        "@enums": "./src/enums",
+        "@packets": "./src/packets",
+        "@structs": "./src/structs",
+        "@managers": "./src/managers",
+        "@core": "./src"
+      }
+    }]
+  ]
+}
