+++ .prettierrc
@@ -9,5 +9,10 @@
   "bracketSpacing": true,
   "arrowParens": "always",
   "proseWrap": "always",
-  "endOfLine": "lf"
+  "endOfLine": "lf",
+
+  "importOrder": ["^[./]", "@(.*)$"],
+  "importOrderSeparation": true,
+  "importOrderSortSpecifiers": true,
+  "importOrderGroupNamespaceSpecifiers": true
 }
