+++ package.json
@@ -10,28 +10,38 @@
     "url": "https://pharmakesbad.software"
   },
   "scripts": {
-    "clean": "rimraf -rf dist && rimraf -rf types",
-    "build": "npm run clean && tsc",
-    "lint": "eslint src/"
+    "clean": "rimraf -rf dist",
+    "build:compile": "babel src -d dist --extensions \".ts\" --no-comments --out-file-extension \".min.js\"",
+    "build:types": "tsc && tsconfig-replace-paths -p tsconfig.json",
+    "build": "npm run clean && npm run build:compile && npm run build:types",
+    "lint": "eslint src"
   },
   "types": "types/index.d.ts",
   "repository": "https://github.com/ThePhar/archipelago.js.git",
   "bugs": "https://github.com/ThePhar/archipelago.js/issues",
   "dependencies": {
-    "module-alias": "^2.2.2",
     "websocket": "^1.0.34"
   },
   "devDependencies": {
+    "@babel/cli": "^7.17.10",
+    "@babel/core": "^7.18.5",
+    "@babel/plugin-transform-typescript": "^7.18.4",
+    "@babel/preset-env": "^7.18.2",
+    "@babel/preset-typescript": "^7.17.12",
+    "@trivago/prettier-plugin-sort-imports": "^3.2.0",
     "@types/module-alias": "^2.0.1",
     "@types/node": "^17.0.42",
     "@types/websocket": "^1.0.5",
     "@typescript-eslint/eslint-plugin": "^5.27.1",
     "@typescript-eslint/parser": "^5.27.1",
+    "babel-plugin-module-resolver": "^4.1.0",
+    "babel-preset-minify": "^0.5.2",
     "eslint": "^8.17.0",
     "eslint-config-prettier": "^8.5.0",
     "prettier": "^2.6.2",
     "rimraf": "^3.0.2",
     "ts-node": "^10.8.1",
+    "tsconfig-replace-paths": "^0.0.11",
     "typescript": "^4.7.3"
   }
 }
