+++ src/index.ts
@@ -1,17 +1,10 @@
-import moduleAlias from "module-alias";
+import { ArchipelagoClient } from "./ArchipelagoClient";
 
-// Set up module aliases.
-moduleAlias.addAliases({
-    "@enums": __dirname + "/enums",
-    "@packets": __dirname + "/packets",
-    "@structs": __dirname + "/structs",
-});
+export * from "@enums";
+export * from "@managers";
+export * from "@packets";
+export * from "@structs";
 
-// Export files from folders.
-export * from "./enums";
-export * from "./packets";
-export * from "./structs";
+export * from "./ArchipelagoClient";
 
-// Set default export.
-import { ArchipelagoClient } from "@structs";
 export default ArchipelagoClient;
