+++ src/packets/BasePacket.ts
@@ -1,5 +1,5 @@
-import { CommandPacketType } from "@enums";
 import * as Packets from "@packets";
+import { CommandPacketType } from "@enums";
 
 export interface BasePacket {
     readonly cmd: CommandPacketType;
